+++ src/render/dom/utils/filter-props.ts
@@ -42,7 +42,9 @@ export function filterProps(
         if (
             shouldForward(key) ||
             (forwardMotionProps === true && isValidMotionProp(key)) ||
-            (!isDom && !isValidMotionProp(key))
+            (!isDom && !isValidMotionProp(key)) ||
+            // If trying to use native HTML drag events, forward drag listeners
+            (props["draggable"] && key.startsWith("onDrag"))
         ) {
             filteredProps[key] = props[key]
         }