+++ src/utils/use-instant-transition-state.ts
@@ -1,0 +1,3 @@
+export const instantAnimationState = {
+    current: false,
+}